moved default adapter
